refactor(perl-ci-hygiene): extract doc path audit into SRP submodule (#0000)
