+++ crates/perl-ci-hygiene/src/commands/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod doc_paths;
